Refactor some class names to support a WebUSB API implementation.
