+++ src/App.tsx
@@ -8,7 +8,7 @@ import Hid from './services/hid/ui/Hid';
 import Top from './components/top/Top.container';
 import KeyboardDefinitionManagement from './components/keyboards/KeyboardDefinitionManagement.container';
 import Catalog from './components/catalog/Catalog.container';
-import { Serial } from './services/serial/ui/Serial';
+import { Firmware } from './services/firmware/ui/Firmware';
 import Documents from './components/documents/Documents.container';
 
 class App extends React.Component<StyledComponentProps, {}> {
@@ -38,7 +38,7 @@ class App extends React.Component<StyledComponentProps, {}> {
         <BrowserRouter>
           <Switch>
             <Route exact path="/hid" component={Hid} />
-            <Route exact path="/serial" component={Serial} />
+            <Route exact path="/firmware" component={Firmware} />
             <Route exact path="/configure" component={Configure} />
             <Route
               exact
