+++ src/services/firmware/Command.ts
