+++ src/services/firmware/FirmwareWriter.ts
@@ -1,4 +1,5 @@
 import { IErrorHandler, IResult } from './Types';
+import { IFirmware } from '../storage/Storage';
 
 export type FirmwareWriterProgressListener = (
   message: string,
@@ -22,18 +23,21 @@ export type FirmwareWriterPhaseListener = (phase: IFirmwareWriterPhase) => void;
 
 export interface IFirmwareWriter {
   read(
+    firmware: IFirmware,
     size: number,
     progress: FirmwareWriterProgressListener,
     phase: FirmwareWriterPhaseListener,
     errorHandler: IErrorHandler
   ): Promise<IFirmwareWriterReadResult>;
   verify(
+    firmware: IFirmware,
     bytes: Uint8Array,
     progress: FirmwareWriterProgressListener,
     phase: FirmwareWriterPhaseListener,
     errorHandler: IErrorHandler
   ): Promise<IResult>;
   write(
+    firmware: IFirmware,
     flashBytes: Uint8Array,
     eepromBytes: Uint8Array | null,
     progress: FirmwareWriterProgressListener,
