+++ src/services/firmware/Types.ts
