+++ src/services/firmware/caterina/CaterinaCommands.ts
