+++ src/services/firmware/ui/Firmware.scss
