+++ src/services/storage/Storage.ts
@@ -2,7 +2,7 @@ import { LayoutOption } from '../../components/configure/keymap/Keymap';
 import { IFirmwareCodePlace, IKeyboardFeatures } from '../../store/state';
 import { IDeviceInformation } from '../hid/Hid';
 import { KeyboardLabelLang } from '../labellang/KeyLabelLangs';
-import { IBootloaderType, IMcuType } from '../serial/Types';
+import { IBootloaderType, IMcuType } from '../firmware/Types';
 
 export interface IResult {
   readonly success: boolean;
